requirements.txt
requests>=2.31.0
pandas>=2.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
schedule>=1.2.0
loguru>=0.7.0
python-dateutil>=2.8.2
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0